SourceD: do not bypass data from or to uncached_get requests.
Originally, we are strictly inclusive.
So in sourceD, we use set and way to check for hazards and bypass data.
But as we make missed get non inclusive, uncached get is not a hit,
and the results may not stay in this way.
So, we can not use uncached_get's way to check for hazards.
Do not bypass data from uncached get.

Also, uncached get should get their data from grantbuffer，if it gets
data from bypass, grantBuffer may not be freed properly.

In conclusion, do not bypass data from or to uncached_get.